requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
google-cloud-storage
python-dotenv
python-magic
python-magic-bin
requests